% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myCLT.R
\name{myclt}
\alias{myclt}
\title{myCLT function}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n, }{number of uniform distributions}

\item{iter, }{number of iterations of n}

\item{a, }{lower bound}

\item{b, }{upper bound}
}
\value{
histogram showing the distribution of the sum of uniforms
}
\description{
myCLT function
}
