% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{mymaxlikg function}
\usage{
mymaxlikg(lfun, theta)
}
\arguments{
\item{lfun, }{object to hold plot}

\item{theta, }{theta value defined by user}
}
\value{
plot graphing max likelihood
}
\description{
mymaxlikg function
}
