% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Curve plotting function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{the mean of the distribution}

\item{sigma}{the standard deviation of the distribution}

\item{a}{the limit for the area covered by the plotting function}
}
\value{
the curve with area shaded and a list with the mean, standard deviation, and
calculated probability
}
\description{
Curve plotting function
}
