% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{mysample function}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{size of experiment}

\item{iter}{number of times to run experiment}

\item{time}{delay on the bar plot creations}
}
\value{
barplots of each iteration of the experiment
}
\description{
mysample function
}
