% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N, }{number of tickets}

\item{gamma, }{gamma value}

\item{p, }{probabilty value}
}
\value{
two graphs showing the optimal tickets for discrete and continuous
distributions
}
\description{
ntickets function
}
